Address some code revew feedback
